Add a "hello-world" action in a module
This is kind of like "MainAction" from real OpenSearch, but we're not
assuming the existence of a cluster and it loads from a plugin.

Since this is our first action and first REST API, it turns out that
there were some support classes missing from the core for convenience
of implementation.

Signed-off-by: Michael Froh <msfroh@apache.org>